+++ nym-vpn-app/src-tauri/src/vpnd/config/vpnd_config.rs
@@ -9,7 +9,7 @@ use crate::vpnd::node::Node;
 use super::mixnet_config::{MixnetTrafficConfig, MixnetTrafficDefaults};
 
 use crate::vpnd::gateway::GatewaySelectionAlgorithmConfig;
-use crate::vpnd::tunnel::{FrontingMode, SplitTunnelSettings};
+use crate::vpnd::tunnel::{FrontingMode, GeoExclusionSettings, SplitTunnelSettings};
 
 #[derive(Serialize, Deserialize, Debug, Clone, TS)]
 #[serde(rename_all = "camelCase")]
@@ -31,6 +31,7 @@ pub struct VpndConfig {
     pub mixnet_traffic: MixnetTrafficConfig,
     pub mixnet_traffic_defaults: MixnetTrafficDefaults,
     pub split_tunnel: SplitTunnelSettings,
+    pub geo_exclusion: GeoExclusionSettings,
     pub gateway_selection_algorithm_config: GatewaySelectionAlgorithmConfig,
     pub gateway_independence_notifications: bool,
 }
@@ -60,6 +61,7 @@ impl VpndConfig {
             mixnet_traffic: config.mixnet_traffic.into(),
             mixnet_traffic_defaults: MixnetTrafficDefaults::get(),
             split_tunnel: config.split_tunnel.into(),
+            geo_exclusion: config.geo_exclusion.into(),
             gateway_selection_algorithm_config: config.gateway_selection_algorithm_config.into(),
             gateway_independence_notifications: config.gateway_independence.enable_notifications,
         })