+++ nym-vpn-app/src/screens/settings/geo-exclusion/components/index.ts
@@ -1,0 +1,2 @@
+export { default as Socks5PortCard } from './Socks5PortCard';
+export { default as ExcludedRegions } from './ExcludedRegions';