+++ nym-vpn-app/src/screens/settings/index.ts
@@ -13,4 +13,5 @@ export * from './mixnet-tuning';
 export * from './account';
 export * from './diagnostic';
 export * from './split-tunneling';
+export * from './geo-exclusion';
 export * from './notifications';